Rename docs/gettingstarted/getting-started.md to docs/get-started.md
standardizing the "get started" path and filename. This is the md file that is linked to the "rocket ship" Get Started. It is specified in tenant.json (getStartedFilePath). That getStartedFilePath needs to point to "docs/get-started.md". In other words, all tenants have to have a docs/get-started.md file and that needs to be referred to by the tenant.json getStartedFilePath